requirements.txt
fastapi
uvicorn[standard]
streamlit
python-dotenv
requests
pydantic
pytest
